misc/HtmlTestSuite/results/test-0538.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Self-Baseline alignment and sizing cyclic dependency</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez Garcia-Boente"/>
<link href="https://drafts.csswg.org/css-grid-1/#alignment" rel="help"/>
<link href="https://drafts.csswg.org/css-grid-1/#column-align" rel="help"/>
<link href="https://drafts.csswg.org/css-grid-1/#row-align" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#propdef-align-items" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#propdef-justify-items" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#baseline-alignment" rel="help"/>
<link href="https://drafts.csswg.org/css-align-3/#valdef-justify-self-baseline" rel="help"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="../../support/alignment.css" rel="stylesheet"/>
<!-- https://github.com/w3c/csswg-drafts/issues/3046 -->
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
.grid {
  position: relative;
  display: inline-grid;
  background: grey;
  text-orientation: sideways;
  font-family: Ahem;
}
.row { grid: minmax(0px, 1fr) / auto 50px 100px }
.column { grid: auto 50px 100px / minmax(0px, 1fr); }
.item1 {
   font-size: 30px;
   background: blue;
}
.item2 {
  font-size: 20px;
  background: red;
}
.item3 {
  font-size: 80px;
  background: green;
}
.height50 { height: 50px; }
.relativeHeight { height: 50%; }
.relativeWidth  { width: 50%; }
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script type="text/javascript">
  setup({ explicit_done: true });
</script>
</head><body onload="document.fonts.ready.then(() =&gt; { checkLayout('.grid'); })">

<!-- NOTE: previously this test was asserting some "cyclic" behaviour where an item would switch baseline alignment participation - this is no longer the case per spec. -->
<pre>flex row</pre>
<div class="grid row alignItemsBaseline">
  <div class="item1 verticalLR" data-expected-height="30" data-expected-width="30" data-offset-x="0" data-offset-y="34">É</div>
  <div class="item2" data-expected-height="20" data-expected-width="50" data-offset-x="30" data-offset-y="48">É</div>
  <div class="item3" data-expected-height="80" data-expected-width="100" data-offset-x="80" data-offset-y="0">É</div>
</div>

<pre>flex row</pre>
<div class="grid row alignItemsBaseline">
  <div class="item1 relativeHeight" data-expected-height="43" data-expected-width="30" data-offset-x="0" data-offset-y="56">É</div>
  <div class="item2" data-expected-height="20" data-expected-width="50" data-offset-x="30" data-offset-y="64">É</div>
  <div class="item3 verticalLR" data-expected-height="80" data-expected-width="100" data-offset-x="80" data-offset-y="0">É</div>
</div>

<pre>flex row</pre>
<div class="grid row alignItemsBaseline">
  <div class="item1 relativeHeight" data-expected-height="40" data-expected-width="30" data-offset-x="0" data-offset-y="40">É</div>
  <div class="item2 verticalLR" data-expected-height="20" data-expected-width="50" data-offset-x="30" data-offset-y="44">É</div>
  <div class="item3" data-expected-height="80" data-expected-width="100" data-offset-x="80" data-offset-y="0">É</div>
</div>

<pre>flex column - all items share a baseline group</pre>
<div class="grid column justifyItemsBaseline">
  <div class="item1" data-expected-height="30" data-expected-width="30" data-offset-x="16" data-offset-y="0">É</div>
  <div class="item2 verticalLR" data-expected-height="50" data-expected-width="20" data-offset-x="12" data-offset-y="30">É</div>
  <div class="item3 verticalLR" data-expected-height="100" data-expected-width="80" data-offset-x="0" data-offset-y="80">É</div>
</div>

<pre>flex column - all items share a baseline group</pre>
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-expected-height="50" data-expected-width="40" data-offset-x="0" data-offset-y="0">É</div>
  <div class="item2 verticalLR" data-expected-height="50" data-expected-width="20" data-offset-x="12" data-offset-y="50">É</div>
  <div class="item3 verticalLR" data-expected-height="100" data-expected-width="80" data-offset-x="0" data-offset-y="100">É</div>
</div>

<pre>flex column - all items share a baseline group</pre>
<div class="grid column justifyItemsBaseline">
  <div class="item1 relativeWidth height50" data-expected-height="50" data-expected-width="40" data-offset-x="0" data-offset-y="0">É</div>
  <div class="item2" data-expected-height="50" data-expected-width="20" data-offset-x="16" data-offset-y="50">É</div>
  <div class="item3 verticalLR" data-expected-height="100" data-expected-width="80" data-offset-x="0" data-offset-y="100">É</div>
</div>

</body>
</html>
